+++ src/sort/omp_sort.cpp
@@ -12,12 +12,12 @@ namespace {
 
 void merge_ranges(vector<int>& arr,
                   vector<int>& scratch,
-                  size_t lo,
-                  size_t mid,
-                  size_t hi) {
-    size_t i = lo;
-    size_t j = mid;
-    size_t k = lo;
+                  long long lo,
+                  long long mid,
+                  long long hi) {
+    long long i = lo;
+    long long j = mid;
+    long long k = lo;
 
     while (i < mid && j < hi) {
         if (arr[i] <= arr[j]) {
@@ -35,17 +35,17 @@ void merge_ranges(vector<int>& arr,
         scratch[k++] = arr[j++];
     }
 
-    for (size_t idx = lo; idx < hi; ++idx) {
-        arr[idx] = scratch[idx];
+    for (long long idx = lo; idx < hi; ++idx) {
+        arr[static_cast<ptrdiff_t>(idx)] = scratch[static_cast<ptrdiff_t>(idx)];
     }
 }
 
 void merge_sort_task(vector<int>& arr,
                      vector<int>& scratch,
-                     size_t lo,
-                     size_t hi,
-                     size_t serial_cutoff) {
-    const size_t len = hi - lo;
+                     long long lo,
+                     long long hi,
+                     long long serial_cutoff) {
+    const long long len = hi - lo;
     if (len <= 1) {
         return;
     }
@@ -56,7 +56,7 @@ void merge_sort_task(vector<int>& arr,
         return;
     }
 
-    const size_t mid = lo + (len / 2);
+    const long long mid = lo + (len / 2);
     const bool spawn_tasks = len >= (serial_cutoff << 1);
 
     #pragma omp task shared(arr, scratch) if(spawn_tasks)
@@ -85,13 +85,17 @@ void omp_merge_sort(vector<int>& arr, int num_threads, int cutoff) {
     }
 
     vector<int> scratch(arr.size());
-    const size_t serial_cutoff = static_cast<size_t>(cutoff);
+    const long long serial_cutoff = static_cast<long long>(cutoff);
 
     omp_set_num_threads(num_threads);
 
     #pragma omp parallel
     {
         #pragma omp single nowait
-        merge_sort_task(arr, scratch, 0, arr.size(), serial_cutoff);
+        merge_sort_task(arr,
+                        scratch,
+                        0LL,
+                        static_cast<long long>(arr.size()),
+                        serial_cutoff);
     }
 }
